Fixing validation for signup page
