+++ praas-app-ionic/src/pages/Signup/index.tsx
@@ -14,9 +14,9 @@ import { useForm } from 'react-hook-form';
 
 import { connect } from 'react-redux';
 import { withRouter, RouteComponentProps } from 'react-router';
-import * as Yup from 'yup';
 import Header from '../../components/Header';
 import { registerUser } from '../../store/user/registration';
+import signupSchema from './schema'
 import './style.scss';
 
 type Props = {
@@ -24,24 +24,12 @@ type Props = {
   registerUser: (data: any) => void;
 };
 
-const signupSchema = Yup.object().shape({
-  firstName: Yup.string()
-    .min(2, 'Must be longer than 2 characters')
-    .max(20, 'Nice try, nobody has a first name that long')
-    .required("Don't be shy. Tell us your first name"),
-  email: Yup.string()
-    .email('Invalid email address')
-    .required('Email is required'),
-  password: Yup.string()
-    .min(2, 'Must be longer than 8 characters')
-    .required('Passphrase is required'),
-});
 const Signup: React.FC<Props & RouteComponentProps> = ({
   user,
   history,
   registerUser,
 }) => {
-  const { register, handleSubmit } = useForm({
+  const { register, handleSubmit, errors } = useForm({
     defaultValues: { firstName: '', email: '', password: '' },
     validationSchema: signupSchema,
   });
@@ -58,7 +46,7 @@ const Signup: React.FC<Props & RouteComponentProps> = ({
     <IonPage className="signup-page">
       <Header />
       <IonContent>
-        <form onSubmit={handleSubmit(onSubmit)}>
+        <form noValidate onSubmit={handleSubmit(onSubmit)}>
           <IonGrid>
             <IonRow className="ion-justify-content-center">
               <IonCol sizeXs="12" sizeSm="4" className="text-align-center">
@@ -66,17 +54,17 @@ const Signup: React.FC<Props & RouteComponentProps> = ({
                   <IonLabel position="floating">Name</IonLabel>
                   <IonInput type="text" name="firstName" ref={register()} />
                 </IonItem>
-                {/* <div className="error">{errors.firstName}</div> */}
+                <div className="error">{errors.firstName?.message}</div>
                 <IonItem>
                   <IonLabel position="floating">Email</IonLabel>
                   <IonInput type="email" name="email" ref={register()} />
                 </IonItem>
-                {/* <div className="error">{errors.email}</div> */}
+                <div className="error">{errors?.email?.message}</div>
                 <IonItem>
                   <IonLabel position="floating">Password</IonLabel>
                   <IonInput type="password" name="password" ref={register()} />
                 </IonItem>
-                {/* <div className="error">{errors.password}</div> */}
+                <div className="error">{errors.password?.message}</div>
                 <IonButton type="submit" color="primary">
                   Submit
                 </IonButton>
